requirements.txt
kivy
pyserial
kivymd
